add cors, validationpipe, interceptor

+++ src/products/products.controller.ts
@@ -1,6 +1,8 @@
-import { Controller, Get, Param, Delete, Post, Body } from '@nestjs/common';
+import { Controller, Get, Param, Delete, Post, Body, Put } from '@nestjs/common';
 import { ProductsService } from './products.service';
 import { CreateProductDTO } from './dtos/create-product.dto';
+import { ParseUUIDPipe, NotFoundException } from '@nestjs/common';
+import { UpdateProductDTO } from './dtos/update-product.dto';
 
 @Controller('products')
 export class ProductsController {
@@ -14,18 +16,34 @@ export class ProductsController {
     };
 
     @Get('/:id')
-    public getById(@Param('id') id: string) {
-        return this.productsService.getById(id);
+    getById(@Param('id', new ParseUUIDPipe()) id: string) {
+      const prod = this.productsService.getById(id);
+      if (!prod) throw new NotFoundException('Product not found');
+      return prod;
     }
 
     @Delete('/:id')
-    public deleteById(@Param('id') id: string) {
-        this.productsService.deleteById(id);
-        return { success: true };
+    deleteById(@Param('id', new ParseUUIDPipe()) id: string) {
+      if (!this.productsService.getById(id))
+        throw new NotFoundException('Product not found');
+      this.productsService.deleteById(id);
+      return { success: true };
     }
 
     @Post('/')
     create(@Body() productData: CreateProductDTO) {
         return this.productsService.create(productData);
     }
+
+    @Put('/:id')
+  update(
+    @Param('id', new ParseUUIDPipe()) id: string,
+    @Body() productData: UpdateProductDTO,
+  ) {
+    if (!this.productsService.getById(id))
+      throw new NotFoundException('Product not found');
+
+    this.productsService.updateById(id, productData);
+    return { success: true };
+  }
 }
